pulled up the custom ResourceBundle.Control class
