+++ kit/src/main/java/com/oracle/javafx/scenebuilder/kit/i18n/I18N.java
@@ -33,12 +33,6 @@ package com.oracle.javafx.scenebuilder.kit.i18n;
 
 import java.text.MessageFormat;
 import java.util.ResourceBundle;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.Locale;
-import java.util.PropertyResourceBundle;
 
 /**
  *
@@ -47,23 +41,7 @@ public class I18N {
     
     private static ResourceBundle bundle;
 
-    private static ResourceBundle.Control UTF8_ENCODING_CONTROL = new ResourceBundle.Control() {
-
-        @Override
-        public ResourceBundle newBundle(String baseName, Locale locale, String format, ClassLoader loader, boolean reload) {
-            String bundleName = toBundleName(baseName, locale);
-            String resourceName = toResourceName(bundleName, "properties");
-
-            try (InputStream is = loader.getResourceAsStream(resourceName);
-                 InputStreamReader isr = new InputStreamReader(is, "UTF-8");
-                 BufferedReader reader = new BufferedReader(isr)) {
-                return new PropertyResourceBundle(reader);
-            } catch (IOException ex) {
-                ex.printStackTrace();
-                return null;
-            }
-        }
-    };
+    private static ResourceBundle.Control utf8EncodingControl = new I18NControl();
     
     public static String getString(String key) {
         return getBundle().getString(key);
@@ -77,7 +55,7 @@ public class I18N {
     public static synchronized ResourceBundle getBundle() {
         if (bundle == null) {
             final String packageName = I18N.class.getPackage().getName();
-            bundle = ResourceBundle.getBundle(packageName + ".SceneBuilderKit"); //NOI18N
+            bundle = ResourceBundle.getBundle(packageName + ".SceneBuilderKit",utf8EncodingControl); //NOI18N
         }
         
         return bundle;
